requirements.txt
pycryptodome==3.21.0
paho-mqtt==1.6.1